Fix behavior of ctrl-a and alt-m.
Atom's editor:move-to-first-character-of-line toggles between start of
line and first character of line if you're already there. This binds
ctrl-a to editor:move-to-beginning-of-line, and makes alt-m behave more
like Emacs, which does not toggle, and just moves straight to the first
nonspace character (or end of line if there's only whitespace).