Bind "ctrl-k ctrl-k" to cut-to-end-of-line on win32.
The built-in win32 keymap has ctrl-k as a prefix key, which creates a 2
second lag before dispatch. This way pressing ctrl-k again will force
the dispatch immediately without the command running twice.

It's not ideal, but an improvement. Might be better to have an optional
way to disable the built-in ctrl-k bindings.

Fixes #75