Fix issue with updating multiple selections while the mark is set.
e.g., if you do the following:

* editor:add-selection-below
* atomic-emacs:forward-char
* atomic-emacs:forward-char
* core:backspace

Then the second selection won't be replaced, as updating the first
selection deactivated the mark, and hence cleared the second selection
before it was updated.

What is needed is to queue the mark for deactivation at a later time.
Here we wait until the end of the current command.