Remove config option to subvert the navigation commands.
This was put in as a hack to workaround an issue that I suspect has been
fixed, as I've never seen it.

As far as I can tell the only difference between the built-in one and
ours is we don't pass the moveToEndOfSelection option, as it does not
make sense with a keyboard-driven mark. Disabling this functionality
results in a crippled atomic-emacs experience, so it's worth attempting
to fix it properly should it happen again.